frontend/requirements.txt
streamlit
pyyaml
python-box